pyproject.toml
[tool.poetry]
name = "aris"
version = "0.1.0"
description = "ARIS: Amplified Reasoning & Intelligence Systems"
authors = ["Laza Bogdan <laza.bogdan@macro-net.ro>"]
license = "MIT"
readme = "README.md"
packages = [{include = "aris"}]

[tool.poetry.scripts]
aris = "aris.cli:main"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
langchain = ">=0.3,<0.4"
langgraph = "^0.3.8"
fastapi = "^0.115.12"
sse-starlette = "^2.1.0"
langchain-openai = "^0.2"
jsonref = "^1.1.0"
datamodel-code-generator = "^0.25.6"
instructor = "^1.3.2"
aiosqlite = "^0.20.0"
hypercorn = "^0.17.3"
python-dotenv = "^1.0.1"
langchain-community = ">=0.3,<0.4"
pypdf2 = "^3.0.1"
chromadb = "^0.5.0"
asyncpg = "^0.29.0"
pydantic-settings = "^2.3.4"
langchain-anthropic = "^0.2"
pygithub = "^2.3.0"
youtube-transcript-api = "^0.6.2"
langchain-experimental = ">=0.3,<0.4"
docker = "^7.1.0"
html2text = "^2024.2.26"
beautifulsoup4 = "^4.12.3"
redis = "^5.0.8"
langsmith = "^0.1.105"
pyairtable = "^2.3.3"
selenium = "^4.24.0"
webdriver-manager = "^4.0.2"
selenium-stealth = "^1.0.6"
sentry-sdk = {extras = ["fastapi"], version = "^2.14.0"}
pyvirtualdisplay = "^3.0"
tldextract = "^5.1.2"
scikit-learn = "^1.5.2"
psycopg = "^3.2.2"
psycopg-pool = "^3.2.3"
langgraph-checkpoint-postgres = "^2.0.16"
google-api-python-client = "^2.145.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
google-cloud-pubsub = "^2.23.1"
gitpython = "^3.1.43"
langchain-chroma = "^0.1.4"
celery = {extras = ["librabbitmq"], version = "^5.4.0"}
gevent = "^24.10.3"
aio-pika = "^9.4.3"
pycryptodome = "^3.21.0"
e2b-code-interpreter = "^1.0.5"
langgraph-prebuilt = "^0.1.2"
composio = "^0.1.1"
prompt-toolkit = "^3.0.51"
mcp = "^1.5.0"
pgai = "^0.10.2"
pgvector = "^0.4.1"
psycopg2-binary = "^2.9.10"
realtimestt = "^0.3"
sounddevice = "^0.4.6"
soundfile = "^0.12.1"
numpy = "^1.26.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.23.8"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
anyio = "^4.8.0"
httpx = "^0.28.0"
# Add other dev dependencies here if needed
freezegun = "^1.5.1"

[tool.pytest.ini_options]
asyncio_mode = "strict"
markers = [
    "integration: marks tests as integration tests (require external services)"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
